[stock-broker-app]-Feature signin

+++ src/models/dbModels.go
@@ -15,7 +15,7 @@ type Users struct {
 	Password       string    `gorm:"column:password" json:"password"`
 	Token          string    `gorm:"column:token" json:"token"`
 	CreationTime   time.Time `gorm:"column:created_at" json:"created_at"`
-	OTP            int       `gorm:"column:otp" json:"otp"`
+	OTP            uint16    `gorm:"column:otp" json:"otp"`
 	EpochTimestamp int64     `gorm:"column:epochtimestamp" json:"epochtimestamp"`
 }
